+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0-rc.1] - 2020-05-17
 ### Changed
 - Changed the internal implementation of how DATIS talks to Windows' TTS (from calling a .NET application to a pure Rust implementation that uses WinRT bindings)